requirements.txt
Flask
requests
beautifulsoup4
python-dotenv
google-generativeai
